AnagramFinder Refactoring Add slf4j & logback
