+++ build.gradle
@@ -11,4 +11,6 @@ repositories {
 
 dependencies {
     testCompile group: 'junit', name: 'junit', version: '4.11'
+    compile 'org.slf4j:slf4j-api:1.7.20'
+    compile 'ch.qos.logback:logback-classic:1.1.7'
 }
