+++ src/main/java/edu/AnagramFinder/AnagramFinder.java
@@ -1,4 +1,4 @@
-package edu;
+package edu.AnagramFinder;
 
 import java.util.HashMap;
