+++ src/main/java/edu/Sorting/MergeSort.java
@@ -1,4 +1,4 @@
-package edu;
+package edu.Sorting;
 
 import java.util.Arrays;
