+++ src/test/java/edu/AnagramFinder/AnagramFinderFastTest.java
@@ -1,4 +1,4 @@
-package edu;
+package edu.AnagramFinder;
 
 import org.junit.Test;
