+++ src/test/java/edu/AnagramFinder/AnagramFinderTest.java
@@ -1,4 +1,4 @@
-package edu;
+package edu.AnagramFinder;
 
 import org.junit.Test;
