+++ src/test/java/edu/Sorting/SortingAlgoTest.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package edu;
+package edu.Sorting;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
